move set scm to application view
