+++ JazminDeployer/src/jazmin/deploy/view/app/ApplicationInfoView.java
@@ -3,21 +3,12 @@
  */
 package jazmin.deploy.view.app;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
-
-import jazmin.core.Jazmin;
-import jazmin.deploy.DeploySystemUI;
-import jazmin.deploy.domain.Application;
-import jazmin.deploy.manager.DeployManager;
-import jazmin.deploy.ui.BeanTable;
-import jazmin.deploy.view.main.CodeEditorCallback;
-import jazmin.deploy.view.main.CodeEditorWindow;
-import jazmin.deploy.view.main.DeployBaseView;
-import jazmin.deploy.view.main.InputWindow;
-import jazmin.deploy.view.main.TaskProgressWindow;
 
 import org.vaadin.aceeditor.AceMode;
 
@@ -32,6 +23,18 @@ import com.vaadin.ui.Label;
 import com.vaadin.ui.TextField;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.themes.ValoTheme;
+
+import jazmin.core.Jazmin;
+import jazmin.deploy.DeploySystemUI;
+import jazmin.deploy.domain.Application;
+import jazmin.deploy.domain.svn.WorkingCopy;
+import jazmin.deploy.manager.DeployManager;
+import jazmin.deploy.ui.BeanTable;
+import jazmin.deploy.view.main.CodeEditorCallback;
+import jazmin.deploy.view.main.CodeEditorWindow;
+import jazmin.deploy.view.main.DeployBaseView;
+import jazmin.deploy.view.main.InputWindow;
+import jazmin.deploy.view.main.TaskProgressWindow;
 
 /**
  * @author yama
@@ -120,9 +123,33 @@ public class ApplicationInfoView extends DeployBaseView{
 		addOptButton("View Template",null, (e)->viewTemplate());
 		addOptButton("System Graph",null, (e)->viewSystemGraph());
 		addOptButton("Instance Graph",null, (e)->viewInstanceGraph());
+		addOptButton("SetScmTag", ValoTheme.BUTTON_PRIMARY, (e) -> setScmTag());
 		//
 		addOptButton("Compile",ValoTheme.BUTTON_PRIMARY, (e)->compileApp());
 	}
+	//
+	
+	private void setScmTag() {
+		List<Application> apps=getOptApps();
+		if(apps.size()==0||apps.size()>1){
+			DeploySystemUI.showInfo("Please select one application");
+			return;
+		}
+		Application app=applications.get(0);
+		if(app.scmPath==null||app.scmPath.isEmpty()){
+			DeploySystemUI.showInfo("This instance scmPath is null");
+			return;
+		}
+		SimpleDateFormat sdf=new SimpleDateFormat("yyyyMMddhhmmss");
+		String targetUrl="svn://web1.itit.io/repo/Tags/"+app.id+sdf.format(new Date());
+		WorkingCopy wc=new WorkingCopy(
+				app.scmUser, 
+				app.scmPassword,
+				app.scmPath,null);
+		wc.copy(targetUrl,"setScmTag");
+		DeploySystemUI.showInfo("set scm tag success targetUrl:"+targetUrl);
+	}
+	//
 	private void viewTemplate0(String appId){
 		String result=DeployManager.getTemplate(appId);
 		if(result==null){
